feat(store): Idx = u64 and slim CI so 0.14 can land on dev
Identity columns store as u64 (retired U = u32); numpy/IO widths are
kept instead of narrowing. WASM domain-uint accessors take
BigUint64Array; JS names stay *U32.

CI: pin rust-toolchain.toml at 1.96.0 (stop floating @stable clippy),
cut the abi3 Python matrix to 3 OS × 3.12, merge fmt into clippy, drop
sccache next to rust-cache, and run rustdoc doctests on the unit job.

+++ molrs-cxxapi/build.rs
@@ -94,7 +94,7 @@ __MOLRS_ELEMENT_VARIANTS__    }
         // readers — owned copies (RefCell precludes returning borrowed slices)
         fn frame_column_f64(fref: &FrameRef, block: &str, col: &str) -> Vec<f64>;
         fn frame_column_i32(fref: &FrameRef, block: &str, col: &str) -> Vec<i32>;
-        fn frame_column_u32(fref: &FrameRef, block: &str, col: &str) -> Vec<u32>;
+        fn frame_column_u32(fref: &FrameRef, block: &str, col: &str) -> Vec<u64>;
         fn frame_column_str(fref: &FrameRef, block: &str, col: &str) -> Vec<String>;
         fn frame_box(fref: &FrameRef) -> Vec<f64>;
 
@@ -115,7 +115,7 @@ __MOLRS_ELEMENT_VARIANTS__    }
             fref: &mut FrameRef,
             block: &str,
             col: &str,
-            data: &[u32],
+            data: &[u64],
         ) -> Result<()>;
         fn frame_set_column_str(
             fref: &mut FrameRef,
